requirements.txt
streamlit==1.28.1
pandas==2.1.3
google-generativeai==0.3.2
python-dotenv==1.0.0
plotly==5.17.0
numpy==1.24.3
openpyxl==3.1.2